Version 0.58: Macros and crash fixes
Macros are now supported. See samples/init.lisp for the current experiement.
Macros are lambdas whose arguments are not evaluated prior to passing, and
which run in the callee's environment.

Fixes a number of instances where the interpreter would crash (segfault) as
a result of not checking argument counts before attempting to use them.